+++ src/ENT3C/__init__.py
@@ -2,5 +2,5 @@ from ENT3C.cli.main import (
     run_get_entropy,
     run_get_similarity,
     run_all,
-    run_locate_largest_euclidean_diff,
+    run_compare_groups,
 )